Refactor extension.ts and add startSqlsClient.ts
