+++ src/extension.ts
@@ -1,29 +1,8 @@
-import { ExtensionContext, Uri, workspace } from "vscode";
-
-import {
-	LanguageClient,
-	LanguageClientOptions,
-	ServerOptions,
-} from "vscode-languageclient/node";
-
-import { delimiter } from "path";
 import * as vscode from "vscode";
-let client: LanguageClient;
+import { client, startSqlsClient } from "./startSqlsClient";
 
-export async function activate(context: ExtensionContext) {
-	const config = parseLanguageServerConfig();
-	const sqlsInPATH = await findSqlsInPath();
-	const serverOptions: ServerOptions = {
-		command: sqlsInPATH.fsPath,
-		args: [...config.flags],
-	};
-
-	const clientOptions: LanguageClientOptions = {
-		documentSelector: [{ language: "sql", pattern: "**/*.sql" }], // TODO: scheme: "file" is needed?
-	};
-
-	client = new LanguageClient("sqls", serverOptions, clientOptions);
-	client.start();
+export async function activate(context: vscode.ExtensionContext) {
+	startSqlsClient();
 
 	// virtual sql files
 	const originalScheme = "sqlsurge";
@@ -80,43 +59,4 @@ export function deactivate(): Thenable<void> | undefined {
 		return undefined;
 	}
 	return client.stop();
-}
-
-interface LanguageServerConfig {
-	flags: string[];
-}
-
-export function parseLanguageServerConfig(): LanguageServerConfig {
-	const sqlsConfig = vscode.workspace.getConfiguration("sqls");
-	const config = {
-		flags: sqlsConfig.languageServerFlags || [],
-	};
-	return config;
-}
-
-async function findSqlsInPath(): Promise<Uri | undefined> {
-	const path = process.env.PATH;
-
-	if (!path) {
-		return;
-	}
-
-	for (const dir of path.split(delimiter)) {
-		const sqls = Uri.joinPath(Uri.file(dir), "sqls");
-		if (await fileExists(sqls)) {
-			return sqls;
-		}
-	}
-}
-
-async function fileExists(path: Uri) {
-	try {
-		await workspace.fs.stat(path);
-		return true;
-	} catch (err) {
-		if (err.code === "ENOENT" || err.code === "FileNotFound") {
-			return false;
-		}
-		throw err;
-	}
 }
